got everything building again post merge

+++ src/Core/Models/Api/Response/ProfileOrganizationResponseModel.cs
@@ -50,7 +50,7 @@ namespace Bit.Core.Models.Api
         public bool UseBusinessPortal => UsePolicies || UseSso; // TODO add events if needed
         public bool UsersGetPremium { get; set; }
         public bool SelfHost { get; set; }
-        public short? Seats { get; set; }
+        public int? Seats { get; set; }
         public short? MaxCollections { get; set; }
         public short? MaxStorageGb { get; set; }
         public string Key { get; set; }
