+++ src/Core/Models/Data/OrganizationUserOrganizationDetails.cs
@@ -19,7 +19,7 @@ namespace Bit.Core.Models.Data
         public bool UseBusinessPortal => UsePolicies || UseSso;
         public bool SelfHost { get; set; }
         public bool UsersGetPremium { get; set; }
-        public short? Seats { get; set; }
+        public int? Seats { get; set; }
         public short? MaxCollections { get; set; }
         public short? MaxStorageGb { get; set; }
         public string Key { get; set; }
